Move template responsibilities out to Spaces, remove createAndAttachComponent from dataModel
